+++ docs/src/dokumentation/RELEASENOTES.md
@@ -7,3 +7,9 @@
 - Build, Test und Deployment Workflows erstellt
 
 ### Verbessert
+
+## Version 2.0 - 14.12.23
+### Hinzugefügt
+- Frontend auf Mobidam angepasst
+- Dummy-Frontend: Schnittstellen Detail-View
+- Zuordnung (Erstellung), Anzeigen und Löschen von Personen zu Schnittstellen in der UI